requirements.txt
tqdm
tensorflow-gpu
numpy
pytables
